pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Sunny Shankar <sunny.shankar44@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.3.4"
fastapi = "^0.111.1"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
uvicorn = "^0.30.3"
redis = "^5.0.7"
requests = "^2.32.3"
sqlmodel = "^0.0.20"
pyjwt = "^2.8.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
